Bgp block on loopback1 testgap
